requirements/base.txt
-e git+https://github.com/randlet/django-genericdropdown.git@473ff52610af659f7d2a3616a6e3322e21673b4d#egg=django_genericdropdown
-e git+https://github.com/randlet/django-recurrence.git@b3a73e8e03952107e58382922fec37aead31fd6f#egg=django-recurrence
-e git+https://github.com/randlet/django-sql-explorer.git@12802fe83f9c45fd0bbe9610cb442dcfc5666d44#egg=django-sql-explorer
black>=23.1.0
Django>=2.2.18
django-q2>=1.3.4
PyVirtualDisplay>=2.0
beautifulsoup4>=4.9.3
concurrent-log-handler>=0.9.19
coreapi>=2.3.3
coverage>=5.4
django-admin-views>=0.8.0
django-auth-adfs>=1.6.0
django-braces>=1.13.0
django-contrib-comments>=1.8.0
django-coverage>=1.2.4
django-crispy-forms>=1.6.0
django-debug-toolbar>=2.0
django-dynamic-raw-id>=2.8
django-extensions>=3.1.0
django-filter>=2.4.0
django-form-utils>=1.0.3
django-formtools>=2.1
django-listable>=0.5.2
django-mptt>=0.10.0
django-mptt-admin>=0.7.2
django-picklefield>=2.0
django-registration>=3.1.1
djangorestframework>=3.12.2
djangorestframework-filters>=1.0.0_dev0
django-widget-tweaks>=1.4.1
freezegun>=0.3.15
html5lib>=1.1
markdown>=3.3.3
matplotlib>=3.3.4
numpy>=1.19.5
openpyxl>=3.0.6
pandas>=1.1.5
pep8>=1.7.1
Pillow>=8.1.0
pydicom>=2.1.2
pylinac-qatrackplus>=2.5.1.3
pynliner>=0.8.0
pytest-cov>=2.11.1
pytest-django>=4.1.0
pytest-sugar>=0.9.4
pytest>=6.2.2
python-dateutil>=2.8.1
pytz>=2021.1
reportlab>=3.5.59
requests<=3
scipy<=1.5.4
selenium>=3.141.0
xlrd>=2.0.1
XlsxWriter>=1.3.7

